feat: upgrade nuxt-graphql-middleware to 5.4.0 and use layer auto-discovery
5.4.0 adds native layer support for graphqlMiddleware.serverOptions.ts and
graphqlMiddleware.clientOptions.ts. Replace the file-copying workaround with
a single layer registration (nuxt.options._layers.push) in the module setup,
so the middleware auto-discovers WPNuxt's defaults from the package root.

- Bump nuxt-graphql-middleware 5.3.2 -> 5.4.0
- Add server/ and app/ to the published package files
- Sync root-level server options with auth cookie handling
- Remove setupServerOptions / setupClientOptions file-copying logic
- Delete 32 duplicate options files across playgrounds, test fixtures, and
  e2e fixtures (all inherit from the WPNuxt layer now)
- Keep playgrounds/core/server/graphqlMiddleware.serverOptions.ts as a live
  demonstration that main-app layer overrides win over the WPNuxt layer
- Fix pre-existing typecheck error in useWPConnection.test.ts

Also picks up 5.4.0's useAsyncGraphqlQuery refetching fix for Nuxt 4.4+
and the corrected operation parameter in server callbacks.

Closes #258
