+++ packages/core/package.json
@@ -40,7 +40,9 @@
   "main": "./dist/module.mjs",
   "types": "./dist/types.d.mts",
   "files": [
-    "dist"
+    "dist",
+    "server",
+    "app"
   ],
   "publishConfig": {
     "access": "public"
@@ -59,7 +61,7 @@
     "defu": "^6.1.4",
     "dompurify": "^3.3.3",
     "graphql": "^16.13.1",
-    "nuxt-graphql-middleware": "5.3.2",
+    "nuxt-graphql-middleware": "5.4.0",
     "scule": "^1.3.0"
   },
   "devDependencies": {
