+++ packages/core/server/graphqlMiddleware.serverOptions.ts
@@ -1,5 +1,6 @@
 import { defineGraphqlServerOptions } from 'nuxt-graphql-middleware/server-options'
-import { getHeader } from 'h3'
+import { getHeader, getCookie } from 'h3'
+import { useRuntimeConfig } from '#imports'
 
 /**
  * WPNuxt default server options for nuxt-graphql-middleware.
@@ -7,18 +8,32 @@ import { getHeader } from 'h3'
  * This enables:
  * - Cookie forwarding for WordPress preview mode
  * - Authorization header forwarding for authenticated requests
+ * - Auth token from cookie for @wpnuxt/auth
  * - Consistent error logging
  *
  * Users can customize by creating their own server/graphqlMiddleware.serverOptions.ts
  */
 export default defineGraphqlServerOptions({
   async serverFetchOptions(event, _operation, _operationName, _context) {
+    // Get auth token from Authorization header or from cookie
+    let authorization = getHeader(event, 'authorization') || ''
+
+    // If no Authorization header, check for auth token in cookie (@wpnuxt/auth)
+    if (!authorization) {
+      const config = (useRuntimeConfig().public as Record<string, unknown>).wpNuxtAuth as { cookieName?: string } | undefined
+      const cookieName = config?.cookieName || 'wpnuxt-auth-token'
+      const authToken = getCookie(event, cookieName)
+      if (authToken) {
+        authorization = `Bearer ${authToken}`
+      }
+    }
+
     return {
       headers: {
         // Forward WordPress auth cookies for previews
         Cookie: getHeader(event, 'cookie') || '',
-        // Forward authorization header if present
-        Authorization: getHeader(event, 'authorization') || ''
+        // Forward authorization header or token from cookie
+        Authorization: authorization
       }
     }
   },
