+++ playgrounds/core/server/graphqlMiddleware.serverOptions.ts
@@ -2,29 +2,26 @@ import { defineGraphqlServerOptions } from 'nuxt-graphql-middleware/server-optio
 import { getHeader } from 'h3'
 
 /**
- * WPNuxt default server options for nuxt-graphql-middleware.
+ * Example: override WPNuxt's default server options.
  *
- * This enables:
- * - Cookie forwarding for WordPress preview mode
- * - Authorization header forwarding for authenticated requests
- * - Consistent error logging
+ * When this file exists in the main app, it takes priority over the defaults
+ * provided by the WPNuxt layer (nuxt-graphql-middleware scans _layers in order,
+ * first match wins). Users can use this pattern to customize request forwarding,
+ * add custom headers, etc.
  *
- * Users can customize by creating their own server/graphqlMiddleware.serverOptions.ts
+ * This playground example reproduces the default cookie/authorization forwarding
+ * and adds an X-Playground header so you can verify the override is active by
+ * inspecting outbound requests to WordPress.
  */
 export default defineGraphqlServerOptions({
-  async serverFetchOptions(event, _operation, _operationName, _context) {
+  async serverFetchOptions(event) {
+    console.log('[LAYER-OVERRIDE-TEST] playground override is being used — main-app layer override works!')
     return {
       headers: {
-        // Forward WordPress auth cookies for previews
-        Cookie: getHeader(event, 'cookie') || '',
-        // Forward authorization header if present
-        Authorization: getHeader(event, 'authorization') || ''
+        'Cookie': getHeader(event, 'cookie') || '',
+        'Authorization': getHeader(event, 'authorization') || '',
+        'X-Playground': 'core'
       }
     }
-  },
-
-  async onServerError(event, error, _operation, operationName) {
-    const url = event.node.req.url || 'unknown'
-    console.error(`[WPNuxt] GraphQL error in ${operationName} (${url}):`, error.message)
   }
 })
